Avoid multiple symbol definitions for MinPvProcessor.
The non-template class MinpvProcessor contains all function
definitions inside of the header file without inlining. The
results in the symbols being in the library of opm-core and
in case of using CpGrid an addtional occurence in other
libraries (opm-autodiff?).

This commit changes this by simply inlining all these functions.

Note that another (better?) option would be to move at least
the bigger once into a cpp file.