Make sibling search for libecl work for newer CMake variants.
I might be wrong, but the old code used to work for CMake <=3.1 (maybe
by accident). It does not for CMake 3.7.1 as macro arguments are not variables
and thus module is treated as string and the condition never gets true.

With this commit we dereference the macro argument module to make the code work.