[bugfix] Make sure to use CONFIG_MODE for subsequent boost searches.
Otherwise boost components found by previous search will be marked as
not found if a new component is not found in the new search.

This happend for #3381 if regex was not found and made cmake believ
that unit_test_framework was not found either. Even though it was in a
previous search.

Now we make sure to CONFIG_MODE if boost was found before, which fixes
this problem. Also we only link dunecommon to test if they are
actually built (read unit_test_framework was marked as found).

Closes #3881