MinpvProcessor: process cells in a vertical column subsequently.
The current implementation might be wrong for cases with multiple
cells in the prospective pinch-out column (e.g. all cells below the
thickness theshold, but the sum is not).
Hence we need to be able to process those cells once only and move kk
index to the next cell not being piched out directly.

It seems like for multiple neighbouring cells begin pinched out the current
algorithm will also add the same NNC multiple times.

For PINCH option 4 will need to be able to compute the harmonic
average of the transmissibilities of the cells beween the active
ones. As those cells are not represented in he grid we will need
to compute the cell centers, face centers, and face areas of the
pinched out cells here and make them available. This might also be
easier with the new loop.